Change train.py for distributed
